virtme-prep-kdir-mods: delete existing $MODDIR/modules.dep at beginning
If preparation process of .virtme_mods is interrupted, the directory
will be in an inconsistent state and needs to updated on the next
run.

Now we use $MODDIR/modules.dep file in run.py to check if virtme_mods
needs to be updated, so we can delete the modules.dep file at the
beginning to ensure that if the command is interrupted, we will update
the directory on the next run.

Signed-off-by: Winston Wen <wentao@uniontech.com>